Player_Tracking/baselines/afl_yolov8_bytrack/requirements.txt
ultralytics==8.3.0
opencv-python>=4.8.0
PyYAML>=6.0

